Fix compiler errors, MSVC
1. static const my_wc_t some_name = whatever;

As far as compiler is concerned, some_name is not a constant,
and can't be used where constant expression is expected, bringing
error C2099: initializer is not a constant
So, either this needs to be replaced with #defines, in C (
in C++, constexpr would be possible)

2. Fix some truncation bugs.

+++ strings/ctype-uca.c
@@ -34408,7 +34408,7 @@ init_weight_level(MY_CHARSET_LOADER *loader, CHARSET_INFO *cs,
   for (i= 0; i != src->contractions.nitems; i++)
   {
     MY_CONTRACTION *item= &src->contractions.item[i];
-    uint length= my_wstrnlen(item->ch, array_elements(item->ch));
+    uint length= (uint)my_wstrnlen(item->ch, array_elements(item->ch));
     uint16 *weights= my_uca_init_one_contraction(&dst->contractions,
                                                  item->ch, length,
                                                  item->with_context);
